requirements.txt
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.8.3
click==8.2.1
colorama==0.4.6
fastapi==0.116.1
flasgger==0.9.7.1
Flask==3.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.2
mistune==3.1.4
packaging==25.0
pluggy==1.6.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.2
pytest-asyncio==1.2.0
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
Werkzeug==3.1.3
